Move the test coordinator back to the core
Pushing the test coordinator to the legacy header breaks client code
that relies on a custom setup with
`test_coordinator_fixture::init_config`.
